backend/requirements.txt
Flask
Flask-Cors
python-docx
PyMuPDF
spacy
openai
werkzeug
fastapi
uvicorn[standard]
jinja2
pdfminer.six
python-multipart
google-generativeai
pydantic>=2
pydantic
pydantic-settings
python-dateutil
jsonpatch
